refactor route students server
