+++ server/test/academy.test.js
@@ -651,111 +651,6 @@ describe('#POST /academy/openCourse', () => {
   });
 });
 
-describe('#GET /academy/coursesOfStudent/:username', () => {
-  let queryCourses;
-  let connect;
-  let username = 'quangnt';
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    queryCourses = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    queryCourses.restore();
-  });
-
-  it('permission denied', (done) => {
-    request(app)
-      .get(`/academy/coursesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        done();
-      });
-  });
-
-  it('failed connect to blockchain', (done) => {
-    connect.returns(null);
-
-    request(app)
-      .get(`/academy/coursesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        username: 'conglt'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        done();
-      });
-  });
-
-  it('failed to query classes of student in chaincode', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
-    });
-
-    let data = JSON.stringify({
-      error: 'Error Network'
-    });
-
-    queryCourses.returns({
-      success: false,
-      msg: data
-    });
-
-    request(app)
-      .get(`/academy/coursesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        username: 'conglt'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        done();
-      });
-  });
-
-  it('success query classes of user student', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
-    });
-
-    let data = JSON.stringify([
-      {
-        CourseID: '1234-456-789-123a',
-        CourseCode: 'ETH101',
-        ShortDescription: 'Ethereum',
-        Description: 'Ethereum101'
-      }
-    ]);
-
-    queryCourses.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .get(`/academy/coursesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        username: 'conglt'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-});
-
 describe('#POST /academy/:subjectId/class', () => {
   let connect;
   let query;
@@ -1051,129 +946,6 @@ describe('#PUT /class', () => {
   });
 });
 
-describe('#GET /academy/classesOfStudent/:username', () => {
-  let queryClasses;
-  let connect;
-  let username = 'quangnt';
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    queryClasses = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    queryClasses.restore();
-  });
-
-  it('permission denied', (done) => {
-    request(app)
-      .get(`/academy/classesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .send({
-        username: 'quangnt'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        done();
-      });
-  });
-
-  it('failed connect to blockchain', (done) => {
-    connect.returns(null);
-
-    request(app)
-      .get(`/academy/classesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        username: 'quangnt'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        done();
-      });
-  });
-
-  it('failed to query classes of student in chaincode', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
-    });
-
-    let data = JSON.stringify({
-      error: 'Error Network'
-    });
-
-    queryClasses.returns({
-      success: false,
-      msg: data
-    });
-
-    request(app)
-      .get(`/academy/classesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        username: 'quangnt'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        done();
-      });
-  });
-
-  it('success query classes of user student', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
-    });
-
-    let data = JSON.stringify([
-      {
-        ClassID: '4ca7fc39-7523-424d-984e-87ea590dfc98',
-        ClassCode: 'Class02',
-        Room: 'Room02',
-        Time: 'Time',
-        Status: 'Register Open',
-        ShortDescription: 'aaaa',
-        Description: 'bbbb',
-        Students: ['student1', 'student2'],
-        TeacherUsername: 'teacher01'
-      },
-      {
-        ClassID: '4ca7fc39-7523-424d-984e-87ea590dfc98',
-        ClassCode: 'Class02',
-        Room: 'Room02',
-        Time: 'Time',
-        Status: 'Register Open',
-        ShortDescription: 'aaaa',
-        Description: 'bbbb',
-        Students: ['student1', 'student2'],
-        TeacherUsername: 'teacher01'
-      }
-    ]);
-
-    queryClasses.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .get(`/academy/classesOfStudent/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        username: 'quangnt'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-});
 describe('#POST /academy/subject', () => {
   let connect;
   let createSubjectStub;
@@ -3010,112 +2782,6 @@ describe('#POST /academy/teacher', () => {
       .then((res) => {
         expect(res.status).equal(500);
         expect(res.body.success).equal(false);
-        done();
-      });
-  });
-});
-describe('#GET /academy/student/:username', () => {
-  let connect;
-  let query;
-  let username = 'quangnt';
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-  });
-
-  it('Can not connect to network', (done) => {
-    connect.returns(null);
-
-    request(app)
-      .get(`/academy/student/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({
-      success: false,
-      msg: 'Can not query chaincode!'
-    });
-
-    request(app)
-      .get(`/academy/student/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-  it('do not success query teacher with student', (done) => {
-    request(app)
-      .get(`/academy/student/${username}`)
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .send({
-        username: 'thienthangaycanh',
-        fullname: 'Tan Trinh'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-  it('success query student of class with admin', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      Username: 'abc',
-      Fullname: 'abcdef',
-      Info: {
-        PhoneNumber: 0376724057,
-        Email: 'abc@gmail.com',
-        Address: 'HN',
-        Sex: 'Male',
-        Birthday: 22 / 03 / 1990,
-        Avatar: 'https://images.com',
-        Country: 'VN'
-      },
-      Courses: 'Blockchain',
-      Classes: 'ClassA'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .get(`/academy/student/${username}`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
         done();
       });
   });
