+++ server/test/students.test.js
@@ -1,0 +1,416 @@
+process.env.NODE_ENV = 'test';
+
+const expect = require('chai').expect;
+const request = require('supertest');
+const sinon = require('sinon');
+const network = require('../fabric/network');
+const app = require('../app');
+const USER_ROLES = require('../configs/constant').USER_ROLES;
+
+require('dotenv').config();
+
+describe('GET /students', () => {
+  let connect;
+  let query;
+
+  beforeEach(() => {
+    connect = sinon.stub(network, 'connectToNetwork');
+    query = sinon.stub(network, 'query');
+  });
+
+  afterEach(() => {
+    connect.restore();
+    query.restore();
+  });
+
+  it('permission denied with student', (done) => {
+    request(app)
+      .get('/students')
+      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(403);
+        expect(res.body.msg).equal('Permission Denied');
+        done();
+      });
+  });
+
+  it('permission denied with teacher', (done) => {
+    request(app)
+      .get('/students')
+      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(403);
+        expect(res.body.msg).equal('Permission Denied');
+        done();
+      });
+  });
+
+  it('Failed to connect blockchain', (done) => {
+    connect.returns(null);
+    request(app)
+      .get('/students')
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(500);
+        done();
+      });
+  });
+
+  it('do not success because query student failed ', (done) => {
+    connect.returns({ error: null });
+
+    query.returns({
+      success: false,
+      msg: 'Query student failed'
+    });
+    request(app)
+      .get('/students')
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(404);
+        done();
+      });
+  });
+
+  it('success query all student with admin academy', (done) => {
+    connect.returns({ error: null });
+    let data = JSON.stringify({ username: 'tantv' }, { username: 'nghianv' });
+
+    query.returns({
+      success: true,
+      msg: data
+    });
+    request(app)
+      .get('/students')
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(200);
+        done();
+      });
+  });
+});
+
+describe('#GET /students/:username', () => {
+  let connect;
+  let query;
+  let username = 'quangnt';
+
+  beforeEach(() => {
+    connect = sinon.stub(network, 'connectToNetwork');
+    query = sinon.stub(network, 'query');
+  });
+
+  afterEach(() => {
+    connect.restore();
+    query.restore();
+  });
+
+  it('Can not connect to network', (done) => {
+    connect.returns(null);
+
+    request(app)
+      .get(`/students/${username}`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(500);
+        expect(res.body.msg).equal('Failed connect to blockchain');
+        done();
+      });
+  });
+
+  it('Can not query chaincode', (done) => {
+    connect.returns({
+      contract: 'academy',
+      network: 'certificatechannel',
+      gateway: 'gateway',
+      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
+    });
+
+    query.returns({
+      success: false,
+      msg: 'Can not query chaincode!'
+    });
+
+    request(app)
+      .get(`/students/${username}`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(404);
+        done();
+      });
+  });
+
+  it('permission denied with role student', (done) => {
+    request(app)
+      .get(`/students/${username}`)
+      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
+      .send({
+        username: 'thienthangaycanh',
+        fullname: 'Tan Trinh'
+      })
+      .then((res) => {
+        expect(res.status).equal(403);
+        expect(res.body.msg).equal('Permission Denied');
+        done();
+      });
+  });
+
+  it('success query student with admin academy', (done) => {
+    connect.returns({
+      contract: 'academy',
+      network: 'certificatechannel',
+      gateway: 'gateway',
+      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
+    });
+
+    let data = JSON.stringify({
+      Username: 'abc',
+      Fullname: 'abcdef',
+      Info: {
+        PhoneNumber: 0376724057,
+        Email: 'abc@gmail.com',
+        Address: 'HN',
+        Sex: 'Male',
+        Birthday: 22 / 03 / 1990,
+        Avatar: 'https://images.com',
+        Country: 'VN'
+      },
+      Courses: 'Blockchain',
+      Classes: 'ClassA'
+    });
+
+    query.returns({
+      success: true,
+      msg: data
+    });
+
+    request(app)
+      .get(`/students/${username}`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(200);
+        done();
+      });
+  });
+});
+
+describe('#GET /students/:username/courses', () => {
+  let queryCourses;
+  let connect;
+  let username = 'quangnt';
+
+  beforeEach(() => {
+    connect = sinon.stub(network, 'connectToNetwork');
+    queryCourses = sinon.stub(network, 'query');
+  });
+
+  afterEach(() => {
+    connect.restore();
+    queryCourses.restore();
+  });
+
+  it('Permission denied with student', (done) => {
+    request(app)
+      .get(`/students/${username}/courses`)
+      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(403);
+        done();
+      });
+  });
+
+  it('Permission denied with teacher', (done) => {
+    request(app)
+      .get(`/students/${username}/courses`)
+      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(403);
+        done();
+      });
+  });
+
+  it('Failed connect to blockchain', (done) => {
+    connect.returns(null);
+
+    request(app)
+      .get(`/students/${username}/courses`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(500);
+        done();
+      });
+  });
+
+  it('Failed to query classes of student in chaincode', (done) => {
+    connect.returns({
+      contract: 'academy',
+      network: 'certificatechannel',
+      gateway: 'gateway',
+      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
+    });
+
+    queryCourses.returns({
+      success: false,
+      msg: 'Query chaincode failed'
+    });
+
+    request(app)
+      .get(`/students/${username}/courses`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(404);
+        done();
+      });
+  });
+
+  it('success query courses of student', (done) => {
+    connect.returns({
+      contract: 'academy',
+      network: 'certificatechannel',
+      gateway: 'gateway',
+      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
+    });
+
+    let data = JSON.stringify([
+      {
+        CourseID: '1234-456-789-123a',
+        CourseCode: 'ETH101',
+        ShortDescription: 'Ethereum',
+        Description: 'Ethereum101'
+      }
+    ]);
+
+    queryCourses.returns({
+      success: true,
+      msg: data
+    });
+
+    request(app)
+      .get(`/students/${username}/courses`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .send({
+        username: 'conglt'
+      })
+      .then((res) => {
+        expect(res.status).equal(200);
+        done();
+      });
+  });
+});
+
+describe('#GET /students/:username/classes', () => {
+  let queryClasses;
+  let connect;
+  let username = 'quangnt';
+
+  beforeEach(() => {
+    connect = sinon.stub(network, 'connectToNetwork');
+    queryClasses = sinon.stub(network, 'query');
+  });
+
+  afterEach(() => {
+    connect.restore();
+    queryClasses.restore();
+  });
+
+  it('Permission denied with student', (done) => {
+    request(app)
+      .get(`/students/${username}/classes`)
+      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(403);
+        done();
+      });
+  });
+
+  it('Permission denied with teacher', (done) => {
+    request(app)
+      .get(`/students/${username}/classes`)
+      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(403);
+        done();
+      });
+  });
+
+  it('Failed connect to blockchain', (done) => {
+    connect.returns(null);
+
+    request(app)
+      .get(`/students/${username}/classes`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(500);
+        done();
+      });
+  });
+
+  it('Failed to query classes of student in chaincode', (done) => {
+    connect.returns({
+      contract: 'academy',
+      network: 'certificatechannel',
+      gateway: 'gateway',
+      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
+    });
+
+    queryClasses.returns({
+      success: false,
+      msg: 'Query classes failed'
+    });
+
+    request(app)
+      .get(`/students/${username}/classes`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(404);
+        done();
+      });
+  });
+
+  it('success query classes of user student', (done) => {
+    connect.returns({
+      contract: 'academy',
+      network: 'certificatechannel',
+      gateway: 'gateway',
+      user: { username: 'hoangdd', role: USER_ROLES.STUDENT }
+    });
+
+    let data = JSON.stringify([
+      {
+        ClassID: '4ca7fc39-7523-424d-984e-87ea590dfc98',
+        ClassCode: 'Class02',
+        Room: 'Room02',
+        Time: 'Time',
+        Status: 'Register Open',
+        ShortDescription: 'aaaa',
+        Description: 'bbbb',
+        Students: ['student1', 'student2'],
+        TeacherUsername: 'teacher01'
+      },
+      {
+        ClassID: '4ca7fc39-7523-424d-984e-87ea590dfc98',
+        ClassCode: 'Class02',
+        Room: 'Room02',
+        Time: 'Time',
+        Status: 'Register Open',
+        ShortDescription: 'aaaa',
+        Description: 'bbbb',
+        Students: ['student1', 'student2'],
+        TeacherUsername: 'teacher01'
+      }
+    ]);
+
+    queryClasses.returns({
+      success: true,
+      msg: data
+    });
+
+    request(app)
+      .get(`/students/${username}/classes`)
+      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
+      .then((res) => {
+        expect(res.status).equal(200);
+        done();
+      });
+  });
+});
